backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.6.0+cu124
langchain==0.3.7
langchain-community==0.3.7
langchain-chroma==0.1.4
langchain-ollama==0.2.3
langgraph==0.2.70
langchain-text-splitters==0.3.2
langchain-huggingface==0.1.2
chromadb==0.5.20
sentence-transformers==3.3.1
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.3
python-multipart==0.0.18
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.28.0
fastapi-mcp==0.3.4
# --- Document processing ---
pdfplumber==0.11.4
pdf2image==1.17.0
pytesseract==0.3.13
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2
pandas==2.2.3
beautifulsoup4==4.12.3
Pillow==11.0.0
python-magic==0.4.27
email-validator==2.2.0
watchfiles==1.0.0
docling==2.31.0
